requirements.txt
pyyaml
requests
fdb
pyodbc
pydantic